LayoutTests/imported/w3c/web-platform-tests/shadow-dom/declarative/tentative/shadowrootadoptedstylesheets/shadowrootadoptedstylesheets-async-fetch-duplicate-bare-specifier-expected.txt


FAIL Bare specifier and resolved URL are not deduplicated but share the same CSSStyleSheet. assert_equals: Bare specifier and its resolved URL are distinct tokens — ordered set parsing should not deduplicate them. expected 2 but got 0
